Correct blots.
